Refactoring of XamlResourceReader and PageNameHelper - Added Services folder - Renamed PageNameHelper.cs to CodeSampleResolver.cs - Renamed IPageNameHelper.cs to ICodeSampleResolver.cs - Removed IXamlResourceReader.cs and XamlResourceReader.cs - Updated references to CodeSampleResolver fixes #51
