--- demo/ToolkitDemo/ToolkitDemo/Helpers/IPageNameHelper.cs
@@ -1,9 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace ToolkitDemo.Helpers
-{
-    public interface IPageNameHelper
-    {
-        IEnumerable<string> GetPageFilesName(string pageName);
-    }
-}
