--- demo/ToolkitDemo/ToolkitDemo/Helpers/IXamlResourceReader.cs
@@ -1,7 +1,0 @@
-﻿namespace ToolkitDemo.Helpers
-{
-    public interface IXamlResourceReader
-    {
-        string ReadEmbeddedResource(string resourceName);
-    }
-}
