+++ demo/ToolkitDemo/ToolkitDemo/Services/ICodeSampleResolver.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace ToolkitDemo.Services
+{
+    public interface ICodeSampleResolver
+    {
+        IEnumerable<string> GetPageFilesName(string pageName);
+        string ReadEmbeddedResource(string resourceName);
+    }
+}
