+++ demo/ToolkitDemo/ToolkitDemo/ToolkitDemo.csproj
@@ -33,4 +33,8 @@
       <DependentUpon>MaterialDatePickerPage.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Helpers\" />
+  </ItemGroup>
 </Project>
